Refactor TSaslTransport
Move the reusable concepts (NegotiationStatus, TSaslServerDefinition)
into a dedicated package.
